Refining

+++ scala_hosting_python/src/main/scala/com/knowscieng/jepp/applydynamic/PythonAge.scala
@@ -2,7 +2,7 @@ package com.knowscieng.jepp.applydynamic
 
 object PythonAge extends App {
   val pythonUtils = new PythonUtils
-  val pythonAge: Float = pythonUtils.py_python_age(100.0)
+  val pythonAge: Double = pythonUtils.py_python_age(100.0)
   println(pythonAge.round)
-  
+
 }
